requirements.txt
amqp==5.0.6
asgiref==3.3.4
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-3-jet==1.0.8
django-celery-beat==2.2.1
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-countries==7.2.1
django-cuser==2017.3.16
django-filter==2.4.0
django-jet==1.0.8
django-logging-middleware==0.1.2
django-modeltranslation==0.17.3
django-request-logging==0.7.3
django-taggit==1.4.0
django-timezone-field==4.2.1
django-tinymce==3.3.0
django-utils-six==2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
image==1.5.33
kombu==5.1.0
loguru==0.5.3
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-resize-image==1.1.19
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
